Rename KafkaClusterRef to KafkaService and refactor responsibilities
Amends with the results of #59

This enables us to configure the TLS options for both inline
bootstrapServers, and kafka connections derived from other resources,
like a reference to strimzi Kafka Custom Resource. Also when we declare
the target cluster topology it will apply to both mechanisms.

KafkaClusterRef was also a painful name as we have adopted `*Ref` as
our property naming convention, pushing us towards a `clusterRefRef`
property. We think KafkaService is a better description.

Signed-off-by: Robert Young <robeyoun@redhat.com>